express validator in route & err format on client
