+++ src/client/js/util/apiNotification.js
@@ -2,10 +2,27 @@ import * as toastr from 'toastr';
 
 const logger = require('@alias/logger')('growi');
 
-export const apiErrorHandler = (err, header = 'Error') => {
-  logger.error(err);
+const errorFormatter = {};
 
-  toastr.error(err, header, {
+errorFormatter.logger = (err) => {
+  return err.toString().replace(/\n/g, ', ');
+};
+
+errorFormatter.toastr = (err) => {
+  const prefixRegexp = /^Error: /;
+  let message = err.toString().replace(/\n/g, '<br>');
+
+  if (prefixRegexp.test(message)) {
+    message = message.replace(prefixRegexp, '');
+  }
+
+  return message;
+};
+
+export const apiErrorHandler = (err, header = 'Error') => {
+  logger.error(errorFormatter.logger(err));
+
+  toastr.error(errorFormatter.toastr(err), header, {
     closeButton: true,
     progressBar: true,
     newestOnTop: false,
@@ -15,7 +32,7 @@ export const apiErrorHandler = (err, header = 'Error') => {
   });
 };
 
-export const apiSuccessHandler = (body, header = '') => {
+export const apiSuccessHandler = (body, header = 'Success') => {
   toastr.success(body, header, {
     closeButton: true,
     progressBar: true,
