--- src/server/form/apiv3/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  userGroup: require('./user-group'),
-};
